Use client objects instead of pointers
This commit fixes an issue when running with the latest go and paho
versions where all the functions that took a client would return an
error like: "type *mqtt.Client is pointer to interface, not interface".
To address this all the functions are updated to take the MQTT.Client
objects directly instead of a pointer.