Remove AsyncDisconnect loop
This commit removes the Async disconnect loop and makes it a simple sync
delete loop. This is necessary to fix (at least on my local system) the
benchmark because close() was getting called on the socket > 1 time by
paho in the async case. I believe this is because of internal changes in
paho.

Since the async component was added because the delete takes can
take some time. (at least according to the comment) But, realistically
the most it will take is 10ms (based on the hardcoded parameter to
Disconnect()) and the time sensitive piece of the benchmarking is
complete there is no reason we can't just wait for each delete to
finish. Especially if that means things won't crash.